Process: Fix sign-compare errors
DoProcess:
oldlen here is defined as a signed integer, with the lenp coming through
as a size_t (unsigned). Convert the local storage to size_t to match, and
allow the compiler to catch any further signed errors later

suFin:
Another len vs local store of len which hasn't been converted to size_t
In this use case, the local l takes the size of the char arrays in
struct inputsu, which are > 0, so this it should be reasonable
to set l as size_t.

+++ src/process.c
@@ -757,7 +757,7 @@ void ProcessInput2(char *ibuf, int ilen)
 
 void DoProcess(Window *window, char **bufp, size_t *lenp, struct paster *pa)
 {
-	int oldlen;
+	size_t oldlen;
 	Display *d = display;
 
 	/* XXX -> PasteStart */
@@ -5226,7 +5226,7 @@ static void suFin(char *buf, size_t len, void *data)
 {
 	struct inputsu *i = (struct inputsu *)data;
 	char *p;
-	int l;
+	size_t l;
 
 	if (!*i->name) {
 		p = i->name;
